Authentication and authorization mechanism implemented.

+++ api/v1/comments/endpoints.py
@@ -8,6 +8,7 @@ import db
 from db.models import Post, Comment
 from . import helpers, schemas
 from .schemas import DisplayComment, DisplayCommentWithPost
+from ..auth.jwt import get_current_user
 
 router = APIRouter()
 
@@ -17,17 +18,21 @@ router = APIRouter()
 # TODO: Add configurable pagination
 # User needs to be logged in for this endpoint to function
 @router.get('/comments/', response_model=Page[DisplayCommentWithPost])
-async def get_user_comments(params: Params = Depends(), database: Session = Depends(db.connection.get_db)):
-    user_id = 9
-    print(user_id)
-    comments = await helpers.get_user_comments(user_id, database)
+async def get_user_comments(params: Params = Depends(),
+                            database: Session = Depends(db.connection.get_db),
+                            current_user=Depends(get_current_user)):
+    current_user_id = int(current_user['sub'])
+    print(current_user_id)
+    print(current_user_id)
+    comments = await helpers.get_user_comments(current_user_id, database)
     return paginate(comments, params)
 
 
 # TODO: Retrieve all comments for post with specific ID
 # TODO: Add configurable pagination
 @router.get('/comments/{id_post}', response_model=Page[DisplayComment])
-async def get_post_comments(id_post: int, params: Params = Depends(),
+async def get_post_comments(id_post: int,
+                            params: Params = Depends(),
                             database: Session = Depends(db.connection.get_db)):
     post = database.query(Post).filter(Post.id == id_post).first()
 
@@ -58,15 +63,17 @@ async def get_comment_by_id(id_comment: int, database: Session = Depends(db.conn
 @router.post('/comments/{id_post}', status_code=status.HTTP_201_CREATED)
 async def create_post_comment(id_post: int,
                               data: schemas.BaseComment,
-                              database: Session = Depends(db.connection.get_db)):
+                              database: Session = Depends(db.connection.get_db),
+                              current_user=Depends(get_current_user)):
+    current_user_id = int(current_user['sub'])
+    print(current_user_id)
     post = database.query(Post).filter(Post.id == id_post).first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id {id_post} does not exist")
 
-    comment_to_create = await helpers.create_comment(data, post, 9,
-                                                     database)  # change user_id to be retrieved from current logged user
+    comment_to_create = await helpers.create_comment(data, post, current_user_id, database)
 
     return comment_to_create
 
@@ -76,27 +83,40 @@ async def create_post_comment(id_post: int,
 @router.put('/comments/{id_comment}', status_code=status.HTTP_200_OK)
 def update_post_comment(id_comment: int,
                         data: schemas.CommentUpdate,
-                        database: Session = Depends(db.connection.get_db)):
-    # dodati provjeru korisnika
+                        database: Session = Depends(db.connection.get_db),
+                        current_user=Depends(get_current_user)):
+    current_user_id = int(current_user['sub'])
+    print(current_user_id)
 
     comment = database.query(Comment).filter(Comment.id == id_comment).first()
 
     if not comment:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Comment with id {id_comment} does not exist")
+
+    if current_user_id != comment.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Comment can only be updated by the user who created it")
 
     updated_comment = helpers.update_comment(comment, data, database)
     return updated_comment
 
 
-# TODO: Update comment for the post with specific ID
+# TODO: Delete comment for the post with specific ID
 # Comment can only be deleted by the user who created it
 @router.delete('/comments/{id_comment}', status_code=status.HTTP_200_OK)
-def delete_post_comment(id_comment: int, database: Session = Depends(db.connection.get_db)):
+def delete_post_comment(id_comment: int,
+                        database: Session = Depends(db.connection.get_db),
+                        current_user = Depends(get_current_user)):
+    current_user_id = int(current_user['sub'])
+    print(current_user_id)
     comment = database.query(Comment).filter(Comment.id == id_comment).first()
 
     if not comment:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Comment with id {id_comment} does not exist")
+    if comment.user_id != current_user_id:
+        raise  HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                             detail="Comment can only be deleted by the user who created it")
     database.delete(comment)
     database.commit()
